gcc so fragile...

+++ wav.c
@@ -69,7 +69,7 @@ tagwav(Tagctx *ctx)
 			csz++;
 			for(n = 0; n < nelem(t); n++){
 				if(memcmp(d, t[n].s, 4) == 0){
-					if(ctx->read(ctx, d, csz) != csz)
+					if((uint)ctx->read(ctx, d, csz) != csz)
 						return -1;
 					d[csz-1] = 0;
 					txtcb(ctx, t[n].type, "", d);
